Compile the test display source only if the CONFIG_CONSOLE_TEST_DISPLAY is enabled in .config
