+++ apps/console_app/test.c
@@ -1,3 +1,5 @@
+#include <board.h>
+
 const unsigned char simons_cat_bin[] = {
   0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
   0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
